Second Brain/Projects/thoughtmap/requirements.txt
# Core
chromadb>=0.5.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Clustering & dimensionality reduction
umap-learn>=0.5.5
hdbscan>=0.8.33

# Utilities
python-Levenshtein>=0.21.0
spacy>=3.7.0

# Optional embedding providers (install as needed)
# pip install openai            # for EMBEDDING_PROVIDER=openai
# pip install google-generativeai  # for EMBEDDING_PROVIDER=google
